typeset -T: fix spurious 2nd error on redef (re: 1fbbeaa1, d309d604)
Reproducers:

    $ ksh -c 'typeset -T foo=(x=1); typeset -T foo=(integer x)'
    ksh: typeset: foo: type cannot be redefined
    ksh: foo: type definition requires compound assignment

    $ ksh -c 'typeset -T foo=(x=1); typeset -T foo=(integer x=2)'
    ksh: foo: type cannot be redefined
    ksh: foo: type definition requires compound assignment

In both cases, the second error message is spurious, as there is in
fact a compound assignment.

The first case was introduced in 1fbbeaa1. The 'integer' command no
longer resolves to a special built-in but is now a regular built-in
so it does not cause the shell to exit on error. The code path
continues and nv_mktype() is called which issues the bad error as
the compound assignment was not fully processed.

The second case was introduced in d309d604. This commit inhibits
the exit-on-error for the assignment list of a regular built-in.
For type definitions, nv_mktype() is called via this code path too.

Since all declaration commands were either special built-ins or
aliases to special built-ins in 93u+, they all exited on error and
this problem never happened. Now that some regular built-ins are
allowed inside 'typeset -T', we need some special-casing to make
them exit on error as before. Arguably this is the correct thing to
do anyway, as their errors are also an error in the enveloping
'typeset' which is a special built-in.

This regression is cosmetic and the case is rare, so this is
probably not worth a NEWS item.

src/cmd/ksh93/sh/xec.c: sh_exec():
- Fix the second case by adding a test for non-NULL 'sh.mktype' to
  the conditions where the assignments list is processed with exit
  on error. This pointer is set while executing a type definition.
- Fix the first case by saving the sh.mktype state before executing
  a built-in and then checking the saved value if a built-in exits
  on error. Saving is needed as sh.mktype has been reset by then.