do not sh_close() a -1 file descriptor (re: 94fc9837, 57cbbce6)
Researching #483 revealed several instances in coprocess cleanup
where sh_close() is called with a file descriptor of -1 (which
flags that the pipe is already closed). As of ec6639f5, this sets
errno to EBADF. While the race condition triggered by this was
fixed in 57cbbce6, it's still better to avoid it.

This re-applies most of the changes reverted in 94fc9837.