Fix assumption that pid_t==int32_t, other wrong typecasts
For $PPID and ${.sh.pid}, ksh has simply been assuming that pid_t
is a 32-bit integer. In POSIX, the size of pid_t is not specified
and operating systems are free to vary it. On any system where
sizeof(pid_t) != 4, these variables produce undefined results.

(Note that this is not a problem with $$ as it is converted to a
string value directly from pid_t in special() in macro.c.)

src/cmd/ksh93/features/options:
- Add feature test that produces NV_PID, a name-value attribute
  macro that is set to NV_INTEGER|NV_SHORT (for int16_t),
  NV_INTEGER (for int32_t) or NV_INTEGER|NV_LONG (for Sflong_t
  a.k.a. intmax_t), depending on the size of pid_t. These are the
  three integer size attributes supported by nv_getval(). The build
  will abort on any (hypothetical?) system where pid_t has a size
  other than these three. In that case, the test and nv_getval()
  will need to have a new integer size attribute added.

src/cmd/ksh93/data/variables.c:
- Use NV_PID instead of NV_INTEGER for $PPID and ${.sh.pid}. This
  fixes the issue.

Other changed files:
- Fix the typecasts to various fmtbase() calls; the first argument
  is of type intmax_t (a.k.a. Sflong_t), not long.
- When outputting PIDs with sfprintf(), typecast to Sflong_t and
  use %lld to account for possible systems with very large PIDs.
  (Though %lld is not in C89/C90, Sfio supports it anyway.)